SP2-265 - Move sentence-plan specific services into their own folder
