+++ server/services/sentence-plan/helloWorld.ts
@@ -1,4 +1,4 @@
-import SentencePlanApiClient from '../data/sentencePlanApiClient'
+import SentencePlanApiClient from '../../data/sentencePlanApiClient'
 
 // TODO: This is a test service and can be removed once we've added other API calls
 export default class HelloWorldService {
